# Replace usage() with freeBytes() in thresholds within hot paths of PoolChunkList