@netudima 2020-03-27
JMH benchmark: https://gist.github.com/netudima/006f0fda91a8f53bc02ca4b825afaacb
Before:
```
Benchmark                              (value)  Mode  Cnt    Score   Error  Units
NettyAllocatorBenchmark.getAndRelease    12345  avgt   15   50.130 ± 0.635  ns/op
NettyAllocatorBenchmark.getAndRelease   123456  avgt   15  114.426 ± 2.168  ns/op
NettyAllocatorBenchmark.getAndRelease  1234567  avgt   15  102.855 ± 2.008  ns/op
```

After:
```
NettyAllocatorBenchmark.getAndRelease    12345  avgt   15   50.308 ± 1.237  ns/op
NettyAllocatorBenchmark.getAndRelease   123456  avgt   15  105.118 ± 1.414  ns/op
NettyAllocatorBenchmark.getAndRelease  1234567  avgt   15   93.763 ± 1.795  ns/op
```
Note: results for buffer with size = 12345 are the same because the alloc/dealloc path for this size does not pass through the changed conditions.