@normanmaurer 2020-03-28
Would it be possible to also check in the benchmark ?