@netudima 2020-03-28
> Would it be possible to also check in the benchmark ?

Yes, I will check how the benchmarks are organized and add it.

One related observation, there are synchronized sections within freeBytes()/usage(). If I got correctly - there were added to get the values as metrics by other threads via PoolChunkMetric API. The question is do we really need these sections in case of access within allocation/deallocation paths? (checking now).

The cost of the synchronized (same bench):
With sync:
```
Benchmark                              (value)  Mode  Cnt   Score   Error  Units
NettyAllocatorBenchmark.getAndRelease   123456  avgt   15  105.118 ± 1.414  ns/op
NettyAllocatorBenchmark.getAndRelease  1234567  avgt   15   93.763 ± 1.795  ns/op
```
Without sync:
```
Benchmark                              (value)  Mode  Cnt   Score   Error  Units
NettyAllocatorBenchmark.getAndRelease   123456  avgt   15  83.440 ± 1.374  ns/op
NettyAllocatorBenchmark.getAndRelease  1234567  avgt   15  76.088 ± 2.419  ns/op
```