@franz1981 2020-03-28
@netudima to correctly evaluation of the synchronization block I suggest to tune the biased lock delay to 0