@netudima 2020-03-28
good point, let me check, at the same time it looks like there are plans to drop biased lock support in future: https://bugs.openjdk.java.net/browse/JDK-8231265