@franz1981 2020-03-28
And maybe (idea for another pr, mine of yours) would be to use a static final const to disable the telemetry of the pool (leaving it true but default) and evaluate the impact by disabling it