@netudima 2020-03-28
note: the same warn regarding biased locking is highlighted in https://mechanical-sympathy.blogspot.com/2011/11/biased-locking-osr-and-benchmarking-fun.html.

Measurements with the delay = 0

with sync in freeBytes() and -XX:BiasedLockingStartupDelay=0:
```
Benchmark                              (value)  Mode  Cnt   Score   Error  Units
NettyAllocatorBenchmark.getAndRelease   123456  avgt   15  83.112 ± 0.489  ns/op
NettyAllocatorBenchmark.getAndRelease  1234567  avgt   15  71.333 ± 1.371  ns/op
```

without sync in freeBytes() and -XX:BiasedLockingStartupDelay=0:
```
Benchmark                              (value)  Mode  Cnt   Score   Error  Units
NettyAllocatorBenchmark.getAndRelease   123456  avgt   15  74.318 ± 1.589  ns/op
NettyAllocatorBenchmark.getAndRelease  1234567  avgt   15  63.752 ± 4.306  ns/op
```
So, yes, there is a significant difference in results compared to the default BiasedLockingStartupDelay: 
```
java -XX:+PrintFlagsFinal -version | grep BiasedLockingStartupDelay
     intx BiasedLockingStartupDelay                 = 4000                                {product}
openjdk version "1.8.0_242"
OpenJDK Runtime Environment (AdoptOpenJDK)(build 1.8.0_242-b08)
OpenJDK 64-Bit Server VM (AdoptOpenJDK)(build 25.242-b08, mixed mode)
```
With biased locking applied the cost of sync is much lower, at the same time it is still not zero.
Note: testing of the option without sync in freeBytes() with and without -XX:BiasedLockingStartupDelay=0 gives different values because there are other sync sections in the allocation/de-allocation paths.