@netudima 2020-03-28
freeBytes field access analysis:

reads:
```
r01 - io.netty.buffer.PoolChunk#usage() - io.netty.buffer.PoolChunk#arena LOCK
r02 - io.netty.buffer.PoolChunk#toString() - io.netty.buffer.PoolChunk#arena LOCK
r03 - io.netty.buffer.PoolChunk#freeBytes() - io.netty.buffer.PoolChunk#arena LOCK
r04     io.netty.buffer.PoolChunkList#allocate() - see w03
r05      io.netty.buffer.PoolChunkList#free() - see w10
r06      io.netty.buffer.PoolChunkList#move() - no lock on arena on this layer
r07        io.netty.buffer.PoolChunkList#move0() - no lock on arena on this layer
r08         io.netty.buffer.PoolChunkList#move() - see r06 (recursion)
r09           io.netty.buffer.PoolChunkList#free() - see w10
r10       io.netty.buffer.PoolChunkList#add() - no lock on arena on this layer
r11         io.netty.buffer.PoolChunkList#add() - see r10 (recursion)
r12         io.netty.buffer.PoolArena#allocateNormal() - see w05
r13         io.netty.buffer.PoolChunkList#allocate() - see w03
```

writes:
```
w01 - io.netty.buffer.PoolChunk#allocateRun() - no lock on arena on this layer
w02     io.netty.buffer.PoolChunk#allocate() - no lock on arena on this layer
w03       io.netty.buffer.PoolChunkList#allocate() - there is a read of freeBytes here, no lock on arena on this layer
w04         io.netty.buffer.PoolArena#allocateNormal() - see w05 
w05       io.netty.buffer.PoolArena#allocateNormal() - no lock on arena on this layer
w06         io.netty.buffer.PoolArena#allocate(...) - LOCK on this arena

w07 - io.netty.buffer.PoolChunk#allocateSubpage() - no lock on arena on this layer
w08     io.netty.buffer.PoolChunk#allocate() - see w03
w09 - io.netty.buffer.PoolChunk#free() - no lock on arena on this layer
w10     io.netty.buffer.PoolChunkList#free() - there is a read of freeBytes here, no lock on arena on this layer
w11       io.netty.buffer.PoolArena#freeChunk() - LOCK on this arena
```
In total, for allocate/de-allocate paths all accesses to freeBytes() are happen from traces protected by Arena lock and sync section can be skipped for these paths.