@franz1981 2020-03-28
I would add a Black hole::consumeCPU with a variable amount of tokens between acquire/release to simulate some cpu activity (on real cases you would allocate-do something-free).