@netudima 2020-03-29
Thank you. Regarding consumeCPU, no problem, I will add it to the bench. 
For my better understanding could you please help to clarify what is the expected effect from it? In case of multi-threaded benchmarks I suppose it may help to make contention between threads more realistic but the current benchmark is a single-threaded. Is it to emulate something like CPU caches eviction?
I found only this comment - https://stackoverflow.com/questions/36288502/jmh-why-do-i-need-blackhole-consumecpu