@franz1981 2020-03-29
@netudima acquire/release on pool involves heavyweight (for the CPU cache) atomic operations for the uncontended case as well (see how an unbiased JDK uncontended locks work too) and consumeCPU is a nice tool to amortize (see  https://shipilev.net/blog/2014/nanotrusting-nanotime/) these operations to study their relative costs (and make possible to compare against other versions that seems to lower such cost). Obviously given that is not such known tool the risk to misuses is high so I am open for discussion to understand if this case is a good candidate or not, but is not mandatory for this PR :)